fix llm json format
